requirements.txt
alembic==1.9.4
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.0.0
fastapi==0.92.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
pytest==7.2.1
pytest-asyncio==0.20.3
python-dateutil==2.8.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.25.0
tomli==2.0.1
typing-extensions==4.5.0
uvicorn==0.20.0
zipp==3.14.0
